Halfway done MagicEngine android fixes
- Added AndroidMacros.h for some quick fixes to save us pain in the future, hopefully
- Removed decrementing ecs iterators cause it isn't supported in clang
- A lot of stupid stuff from clang that we have to work around
- Some actual fixes about libraries not available on android

+++ extern/include/properties/Properties.h
@@ -111,7 +111,7 @@ public:
     constexpr str_view(const char* a, int& i) noexcept
         : m_pStr(a)
         , m_Size
-        { static_cast<std::size_t>
+        { static_cast<uint32_t>
             ([](const char* a, int& i) constexpr
             {
                for (i = 0; a[i] && a[i] != '/' && a[i] != '['; i++);
